shaka player error code 2000
Dear friends, 
I hope you are well. I am facing this problem with shaka Player and the error code 2000. Could you please help me and let me know what I can do to correct this. It is currently working for videos with a DRM protection but not for the contents without any DRM license. 
Thank you.